Update macOS icon to follow Big Sur design
First of all, a greeting to you who will review my contribution, I already know that an icon for Big Sur has already been created and accepted but I wanted to bring my work to your attention because, in my opinion, it is more in line with the Big Sur guidelines.
The icon was created using Apple's official icon template, using the same gradient as the Bitwarden icon for iOS.
Unlike the current icon, the shield is completely white with the addition of internal and external shadows. (I attach the .fig file for further information and / or changes)

https://www.figma.com/file/n5v8SQzg7LKaYXe6ZWtuir/macOS-Big-Sur-Bitwarden-Icon-Template