Add support for setting tray icon in Linux.
The icon used for the tray is the same as the application icon
which has a blue background, this does not match other tray
icons on most Linux desktop themes which are monochromatic.

Electron does not provide an easy way to set an icon that can
be adapted to the current theme in Linux so this allows the user
to pick an icon that matches the theme.

Sample light and dark icons have been provided.